# 2-2-t/specs/001-smartbooking-scheduling/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: SmartBooking API
  version: 0.1.0
  description: |
    Contract-first REST API for SmartBooking.

    Notes:
    - Auth: JWT Bearer token in Authorization header.
    - All errors use a unified envelope with stable error codes and requestId.
servers:
  - url: http://localhost:3001
...
paths:
...
  /api/provider/services:
    get:
      tags: [Provider]
      summary: List services owned by current provider
      operationId: providerListServices
      security:
        - bearerAuth: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/Service'
                required: [items]
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
    post:
      tags: [Provider]
      summary: Create a service
      operationId: providerCreateService
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateServiceRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Service'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/provider/services/{serviceId}:
    patch:
      tags: [Provider]
      summary: Update a service
      operationId: providerUpdateService
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/serviceId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateServiceRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Service'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /api/bookings:
    get:
      tags: [Booking]
      summary: List my bookings
      operationId: listMyBookings
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/BookingStatus'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/Booking'
                required: [items]
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      tags: [Booking]
      summary: Create a booking (reserve a seat)
      operationId: createBooking
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateBookingRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Booking'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
  /api/bookings/{bookingId}/cancel:
    post:
      tags: [Booking]
      summary: Cancel my booking (deadline enforced)
      operationId: cancelBooking
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/bookingId'
      responses:
        '204':
          description: No Content
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    serviceId:
      name: serviceId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    bookingId:
      name: bookingId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
    BadRequest:
      description: Bad Request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Not Found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Conflict:
      description: Conflict
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
...
    CreateServiceRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        description:
          type: string
        durationMinutes:
          type: integer
          minimum: 1
        status:
          type: string
          enum: [ACTIVE, INACTIVE]
      required: [name, description, durationMinutes]
    UpdateServiceRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        description:
          type: string
        durationMinutes:
          type: integer
          minimum: 1
        status:
          type: string
          enum: [ACTIVE, INACTIVE]
      additionalProperties: false
    Service:
      type: object
      properties:
        id:
          type: string
          format: uuid
        providerId:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        durationMinutes:
          type: integer
        status:
          type: string
          enum: [ACTIVE, INACTIVE]
        createdAt:
          type: string
          format: date-time
      required: [id, providerId, name, description, durationMinutes, status, createdAt]
...
    CreateBookingRequest:
      type: object
      properties:
        timeSlotId:
          type: string
          format: uuid
      required: [timeSlotId]
    BookingStatus:
      type: string
      enum: [PENDING, CONFIRMED, CANCELLED, COMPLETED]
    Booking:
      type: object
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        timeSlotId:
          type: string
          format: uuid
        status:
          $ref: '#/components/schemas/BookingStatus'
        createdAt:
          type: string
          format: date-time
        cancelledAt:
          type: string
          format: date-time
          nullable: true
        completedAt:
          type: string
          format: date-time
          nullable: true
      required: [id, userId, timeSlotId, status, createdAt]
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
              description: Stable internal error code
              example: BOOKING_SEAT_FULL
            message:
              type: string
              description: Human-readable error message
            requestId:
              type: string
              description: Correlation id for support/debugging
          required: [code, message, requestId]
      required: [error]